+++ steps/edit-appeal/EditAppeal.js
@@ -2,6 +2,7 @@ const paths = require('paths');
 const { Redirect } = require('@hmcts/one-per-page');
 const { goTo } = require('@hmcts/one-per-page/flow');
 const config = require('config');
+const { resetJourney } = require('middleware/draftAppealStoreMiddleware');
 
 const multipleDraftsEnabled = config.get('features.multipleDraftsEnabled.enabled') === 'true';
 
@@ -12,8 +13,12 @@ class EditAppeal extends Redirect {
 
   handler(req, res, next) {
     if (multipleDraftsEnabled && req.method === 'GET') {
+
       const caseId = req.query.caseId;
+
       if (req.query.caseId && req.session.drafts && req.session.drafts[caseId]) {
+
+        resetJourney(req);
         const draft = req.session.drafts[caseId];
         draft.isUserSessionRestored = true;
         draft.entryPoint = 'Entry';
@@ -22,6 +27,7 @@ class EditAppeal extends Redirect {
       } else {
         res.redirect(paths.session.entry);
       }
+
     } else {
       super.handler(req, res, next);
     }
